# Simplify to_float