* Use right-sized uint type for bit fiddling
  * This removes the need to special casing on endianness
* Replace ternary with just shifting the sign at the right place
  * This seems to improve codegen (less instructions, no cmov)

Codegen demonstration:

gcc 12: https://godbolt.org/z/ceYK4crde
clang 15: https://godbolt.org/z/WTcqn8xGv
MSVC 19.14: https://godbolt.org/z/P3d8dv9Eq