@lemire 2023-03-02
@leni536 Can you run this benchmark in privileged mode (e.g., with sudo) so that you get the performance counters?