@leni536 2023-03-02
> @leni536 Can you run this benchmark in privileged mode (e.g., with sudo) so that you get the performance counters?

# 64 bit

main branch (b1d27734c5443c91847a8fc36e0e32ef23fc9f38):

```
fastfloat                               :  1390.38 MB/s (+/- 0.3 %)    66.27 Mfloat/s      11.73 i/B   258.07 i/f (+/- 0.0 %)      2.88 c/B    63.38 c/f (+/- 0.2 %)      4.07 i/c      4.20 GHz 
```

This PR (https://github.com/leni536/fast_float/commit/6d2fb68f5c928f75642762b235875cfbbd807687):

```
fastfloat                               :  1487.16 MB/s (+/- 0.3 %)    70.88 Mfloat/s      11.82 i/B   260.07 i/f (+/- 0.0 %)      2.69 c/B    59.24 c/f (+/- 0.1 %)      4.39 i/c      4.20 GHz
```

# 32 bit

main branch:

```
fastfloat                               :  1504.07 MB/s (+/- 0.3 %)    71.69 Mfloat/s      11.64 i/B   256.00 i/f (+/- 0.0 %)      2.66 c/B    58.56 c/f (+/- 0.2 %)      4.37 i/c      4.20 GHz
```

This PR:

```
fastfloat                               :  1483.64 MB/s (+/- 0.8 %)    70.71 Mfloat/s      11.73 i/B   258.00 i/f (+/- 0.0 %)      2.70 c/B    59.35 c/f (+/- 0.2 %)      4.35 i/c      4.20 GHz
```

compiler: gcc-12 (Debian 12.2.0-13) 12.2.0
CPU: AMD Ryzen 5 3600